Move stream operator scaffolding
Put StreamFrom, StreamAccept, and StreamTo under tenzir/operator instead of tenzir/async. These templates own operator lifecycle and subpipeline mechanics, while the lower-level stream transport helpers stay in tenzir/async.
