+++ libtenzir/builtins/operators/from_unix_socket.cpp
@@ -7,11 +7,11 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 #include <tenzir/async/stream.hpp>
-#include <tenzir/async/stream_from.hpp>
 #include <tenzir/async/uds.hpp>
 #include <tenzir/compile_ctx.hpp>
 #include <tenzir/file.hpp>
 #include <tenzir/ir.hpp>
+#include <tenzir/operator/stream_from.hpp>
 #include <tenzir/operator_plugin.hpp>
 #include <tenzir/option.hpp>
 #include <tenzir/pipeline_metrics.hpp>
