+++ libtenzir/include/tenzir/operator/stream_accept.hpp
